feat(Comments): implement showComments parameter and normalize line breaks in comments
