+++ shared/src/api/queries/activities/util/activitiesHelpers.ts
@@ -222,8 +222,14 @@ export const taskProvideTags = (result: Task[], type = 'task', entityType = 'tas
       ]
     : [{ type, id: entityType.toUpperCase() + 'S' }]
 
-export const filterKey = (filter: any): string =>
-  typeof filter === 'string' ? filter : JSON.stringify(filter) ?? ''
+export const filterKey = (filter: any): string => {
+  if (typeof filter === 'string') return filter
+  try {
+    return JSON.stringify(filter) ?? ''
+  } catch {
+    return ''
+  }
+}
 
 export const filterActivityTypes: Record<string, string[]> = {
   activity: ['comment', 'version.publish', 'status.change', 'assignee.add', 'assignee.remove'],
