+++ shared/src/containers/ProjectTreeTable/context/ClipboardContext.tsx
@@ -29,6 +29,7 @@ import {
   isSubtasksTSV,
   tsvToSubtasks,
   sanitizeSubtaskName,
+  commentsToText,
 } from './clipboard'
 import { validateClipboardData } from './clipboard/clipboardValidation'
 import { ClipboardContextType, ClipboardProviderProps } from './clipboard/clipboardTypes'
@@ -171,6 +172,10 @@ export const ClipboardProvider: React.FC<ClipboardProviderProps> = ({
               // @ts-ignore
               const subtasks = getCellValue(entity, colId) || []
               cellValue = subtasksToTSV(subtasks, entity.id, entity.name)
+            } else if (colId === 'comments') {
+              // @ts-ignore
+              const comments = getCellValue(entity, 'latestComments') || []
+              cellValue = commentsToText(comments)
             } else {
               // @ts-ignore
               let foundValue = getCellValue(entity, colId)
@@ -204,8 +209,14 @@ export const ClipboardProvider: React.FC<ClipboardProviderProps> = ({
               }
             }
 
-            // Escape double quotes in the cell value and wrap in quotes
-            rowValues.push(`${cellValue.replace(/"/g, '""')}`)
+            // Wrap multi-line / tab / quote values in quotes so spreadsheets keep them in one
+            // cell. Subtasks embed tabs+newlines as real grid structure for paste round-trips —
+            // leave them raw so they stay multi-row/column.
+            const escaped = cellValue.replace(/"/g, '""')
+            const needsQuoting =
+              colId !== 'subtasks' &&
+              (cellValue.includes('\n') || cellValue.includes('\t') || cellValue.includes('"'))
+            rowValues.push(needsQuoting ? `"${escaped}"` : escaped)
           }
 
           // Add row to clipboard text
