+++ shared/src/containers/ProjectTreeTable/context/clipboard/index.ts
@@ -2,3 +2,4 @@ export * from './clipboardTypes'
 export * from './clipboardUtils'
 export * from './clipboardValidation'
 export * from './clipboardSubtasks'
+export * from './clipboardComments'
